Create disk_image.ttl
Note: This example was created for modeling purposes only and is not a reflection of actual data or an existing implementation.

Use case:  Emory University acquired a laptop as a part of the Rushdie papers.  Imaging of the laptop drive was outsourced to a vendor who used GNU ddrescue 1.19.  Following receipt at Emory, a digital archivist reviewed the image, verified the MD5 checksum, and calculated a SHA1 checksum. Ewfacquire was used to migrate the disk image from .dd to .E01 in preparation for ingest into Emory’s digital repository. Individual files (including a Mac sticky note) were extracted from the image using FTK Imager and ingested as separate files.  All ingested files undergo validation, a virus check, and checksum generation as a part of the ingest process.  Users may view files in a emulated environment, which uses SheepShaver 2.3 on a Mac Powerbook.  This emulated environment has been imaged as well and ingested into the repository for preservation purposes.

The disk images are restricted to the public but the sticky note (an example of an extracted file) is available for reading room use only per the deed of gift.